Separate github extraction and file dump
